Update __main__.py
So I added conditional if.args statements so that 1) if --no-ssl flag is set that it will honour this, 2) if neither --no-ssl or --keyfile AND --certfile is set it will default to what @MrAsterico originally intended, 3) if both --keyfile and --certfile are set that it will load the cert and keys using the code you provided earlier. 

Your thoughts on this solution would be useful.